refactor; WIP on pagination
