+++ solvis_graphql_api/composite_solution/__init__.py
@@ -1,0 +1,9 @@
+from .schema import (
+    CompositeRuptureDetail,
+    FilterCompositeSolution,
+    analyse_composite_solution,
+    composite_rupture_detail,
+)
+
+from .composite_solution_schema import CompositeSolutionAnalysisArguments
+from .composite_rupture_detail import CompositeRuptureDetailArguments
